+++ lib/Core.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Mindbox;
+
+trait Core
+{
+    /**
+     * @return Mindbox
+     */
+    public static function mindbox()
+    {
+        return Options::getConfig();
+    }
+}
